Relocate PysuiConfiguraiton
